stt_agent/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
google-cloud-speech>=2.19.0
pydub>=0.25.1
python-dotenv>=1.0.0